Write the simplification oracle; it finds a non-convergence bug
`pop_bubbles` landed in cd4256c against stub aligners and unit fixtures. This
adds the oracle that runs it on real recorded graphs, the top-level entry point
it needs, and the fix for the first thing the oracle found.

`simplify_graph` is the port of `SimplifyGraph.py::simplifyGraph` --- the thin
wrapper the driver is actually called through. `Counting<C>` wraps any
`Consensus` and counts spoa and alignment calls, which is what lets the oracle
say whether a disagreement can be blamed on this port or not.

`rust/tests/simplify_oracle.rs` rebuilds the graph the reference had on entry to
`simplifyGraph` from a `simplify_*.txt` dump (BN/BE/BT records, replayed in
insertion order because node and adjacency order decide `nx.topological_sort`,
which decides which node pairs are candidate bubbles), asserts the rebuild
reproduces the recorded topological order as a precondition, runs the stage with
the real spoa/parasail engine, and diffs against the recorded AN/AE state.
`dump_reference.py` gained `--stage {graph,simplify,both}` and the
BN/BE/BT/AN/AE/AT writers to produce those dumps; bench/README.md documents both
oracles and why they differ in strength. CI now runs both against
bench/corpus/sirv_small.

Because this stage calls spoa, a disagreement is not automatically the port's
fault: `crate::poa` has not been re-verified against the spoa binary on
isONform's own inputs (isONcorrect's 505/505 exact-consensus result does not
transfer --- different sequences, different stage). So the oracle classifies each
disagreeing case by whether spoa was ever called. Zero calls means the
two-supporting-read path, which never touches spoa, diverged: unambiguously a
port bug, and it fails the build. Any calls means the case is reported but does
not fail, pending that verification. A cap hit is always attributable regardless
of spoa calls, since no consensus difference can turn a terminating loop into a
non-terminating one --- the first version of this file got that backwards and
would have hidden exactly the bug below.

Two things came out of pointing it at real data.

Finding 23: read positions go negative after a bubble is popped.
`additional_node_support` invents a position from `prev_end + relative_dist`
then `newend - avg_len`, and both can come out negative: 57 nodes across 16
recorded Drosophila simplifications carry a negative position afterward, where
none did before. `ReadInfo`'s two position fields and `ConsensusAttr` are now
`i64` throughout graph.rs, graph_build.rs and simplify.rs. They were `u32`,
which had been silently clamping these to 0 --- an unmeasured divergence the
oracle exposed the moment it tried to parse a real dump. `py_slice` reproduces
Python's negative-index slicing where the reference does not clamp
(`generate_consensus_path`'s two-read branch, `align_bubble_nodes`'s single-read
branch) as against where it does (the spoa branch clamps `pos1` explicitly).

And the bug worth having built the oracle for: on its first run against 16 real
Drosophila graphs, all 16 disagreed, every one by hitting the iteration cap,
where the reference terminates in 5-9 iterations.

`remove_edges` (SimplifyGraph.py:279-304) records an edge for deletion whenever
it *exists*, unconditionally. The port instead only lifted an edge when
`Graph::edge_length` returned `Some`. Finding 16 already established why that is
the wrong test: an edge `prepare_adding_edges` re-adds after an earlier pop
carries no `length` at all, because `upsert_edge_support` creates it without one.
So the length-based check treated "exists with no length" as "does not exist" and
skipped removing it. That edge then survived every later pop touching its
endpoints, which left its bubble looking poppable for as long as the loop cared
to ask --- the graph a later iteration inspects is corrupted in exactly the way
needed to keep the loop fed.

Fixed by gating on `Graph::has_edge`, and carrying the length through as the
`Option<i64>` the graph already models it as; `LiftedEdge::length` is
`Option<i64>` rather than `i64` so "exists with no length" is representable at
that call site too. Regression test
`a_lengthless_edge_inside_the_bubble_is_still_lifted_and_removed` recreates the
exact shape --- an edge removed then re-added through `upsert_edge_support` ---
and asserts it is both lifted and actually gone from the graph.

Found by replaying one failing dump through the reference in a real
`networkx.DiGraph` and tracing both sides iteration by iteration until a single
bubble's edge walk diverged. Worth recording because the cheaper hypotheses were
all wrong: spoa was returning a valid consensus, the rebuild was byte-exact
against the reference at the end of iteration 1, and an earlier real fix in the
multi-path branch (cloning the two paths where the reference mutates the same
tuples that stay in `all_paths`, so `remove_edges`' `pop(0)` is visible to later
pairs) changed nothing about the cap.

One existing test's premise rested on the same bug without knowing it.
`an_aligner_that_never_refuses_is_stopped_by_the_cap` claimed to document that
the reference's loop has no termination guarantee, using a three-path bubble and
an always-yes aligner that did hit the cap --- because of this bug. With it fixed
the same fixture converges in three iterations, two pops, six edges down to four.
Renamed to `an_aligner_that_never_refuses_still_converges_by_merging_paths_away`
and rewritten to assert that. The structural point stands --- `while
has_combinations` has no termination proof and `hit_iteration_cap` stays as a
safety net (Finding 19) --- but this fixture was not evidence for it, and no
fixture on hand now is.

Verified: all 16 real Drosophila cases agree, spoa called up to 8825 times on
one, none hitting the cap, every rebuild reproducing the reference's topological
order. A wider 56-cluster sample stratified across the corpus's size range
(1.2 KB to 174 KB of reads, largest 5 of 561 excluded to bound runtime) agrees
on 54; the other 2 call spoa 75 and 178 times and so land in the
pending-spoa-verification bucket rather than the cap-hit one, and none of the 56
hit the cap either. sirv_small's two cases still agree with zero spoa calls, so
the CI gate is a real one there. The graph oracle still passes on every recorded
call. 148 unit tests, `clippy --all-targets -D warnings` clean, `fmt --check`
clean.

Closing those 2 remaining cases needs the spoa equivalence check, which is the
next thing owed.

+++ rust/src/graph_build.rs
@@ -148,8 +148,8 @@ pub fn generate_graph_from_intervals(
             t,
             r,
             ReadInfo {
-                start_mini_end: len,
-                end_mini_start: len,
+                start_mini_end: len as i64,
+                end_mini_start: len as i64,
                 original_support: true,
             },
         );
@@ -201,8 +201,8 @@ pub fn generate_graph_from_intervals(
                             prior,
                             r_id,
                             ReadInfo {
-                                start_mini_end: inter.start,
-                                end_mini_start: inter.end,
+                                start_mini_end: inter.start as i64,
+                                end_mini_start: inter.end as i64,
                                 original_support: true,
                             },
                         );
@@ -224,8 +224,8 @@ pub fn generate_graph_from_intervals(
                                 alt,
                                 r_id,
                                 ReadInfo {
-                                    start_mini_end: inter.start,
-                                    end_mini_start: inter.end,
+                                    start_mini_end: inter.start as i64,
+                                    end_mini_start: inter.end as i64,
                                     original_support: true,
                                 },
                             );
@@ -268,8 +268,8 @@ pub fn generate_graph_from_intervals(
                                 prior,
                                 r_id,
                                 ReadInfo {
-                                    start_mini_end: inter.start,
-                                    end_mini_start: inter.end,
+                                    start_mini_end: inter.start as i64,
+                                    end_mini_start: inter.end as i64,
                                     original_support: true,
                                 },
                             );
@@ -305,8 +305,8 @@ pub fn generate_graph_from_intervals(
                                         alt,
                                         r_id,
                                         ReadInfo {
-                                            start_mini_end: inter.start,
-                                            end_mini_start: inter.end,
+                                            start_mini_end: inter.start as i64,
+                                            end_mini_start: inter.end as i64,
                                             original_support: true,
                                         },
                                     );
@@ -343,8 +343,8 @@ pub fn generate_graph_from_intervals(
                                         alt,
                                         r_id,
                                         ReadInfo {
-                                            start_mini_end: inter.start,
-                                            end_mini_start: inter.end,
+                                            start_mini_end: inter.start as i64,
+                                            end_mini_start: inter.end as i64,
                                             original_support: true,
                                         },
                                     );
@@ -370,8 +370,8 @@ pub fn generate_graph_from_intervals(
                         n,
                         r_id,
                         ReadInfo {
-                            start_mini_end: inter.start,
-                            end_mini_start: inter.end,
+                            start_mini_end: inter.start as i64,
+                            end_mini_start: inter.end as i64,
                             original_support: true,
                         },
                     );
@@ -392,8 +392,8 @@ pub fn generate_graph_from_intervals(
                         n,
                         r_id,
                         ReadInfo {
-                            start_mini_end: inter.start,
-                            end_mini_start: inter.end,
+                            start_mini_end: inter.start as i64,
+                            end_mini_start: inter.end as i64,
                             original_support: true,
                         },
                     );
